[Go] Concepts List
Here is a list of Concepts I have started in the early discussions around `v3`. This is by no means a finished list!

I would like to use this PR as a discussion ground. Everything is in question on this list!

My approach for this list was looking through the 3 links mentioned at the bottom and see what features/concepts Go has. Another approach would be to look at an exercise and break down which concepts are needed to solve it. Yet another approach would be to look at other tracks (e.g. C#, JavaScript) and see what concepts they use and whether that applies to Go. For example we do not have the concept of `Strings` in here yet where we could teach students about working with strings (e.g. concatenation, manipulation, iteration, etc.). Maybe that even needs multiple concept exercises.