[Go] Add numbers concept exercise
This is based off of the C# exercise and slightly tweaked to work for
Go. The conditional aspect of the exercise has been stubbed out to keep
the exercise focussed on the core topic and have added more detail about
Type Conversion in Go as is a required part of the exercise.
Also added links to the Go Playground to the examples so students can play
around and explore there if they want to.