[Haskell] Sync v2 and v3 directory structure
This merges the existing [v2 track's directory structure][v2] with
[v3 track's directory structure][v3] as outlined in the v3 repository.

In particular, the layout of multiple example solutions is preserved
until Haskell track contributors come to another conclusion than the
current. This discussion is extracted into exercism/haskell#893.

The ordering of the files in the ASCII drawing is incidentally made so
that the track-specific files are listed first.

[v2]: https://github.com/exercism/haskell#directory-structure
[v3]: https://github.com/exercism/v3/blob/master/docs/maintainers/generic-how-to-implement-a-concept-exercise.md