[Haskell] [WIP] Create concept-exercise.hsfiles template
Based on the directory structure of #542 a Stack template is devised
that will contain the necessary files for creating a concept exercise.

Assuming one is in the directory `languages/haskell/exercises/concept`,
it is possible to type the following:

```
$ stack new foobar ../../concept-exercise.hsfiles -p "camel-name:FooBar"
$ cd foobar
$ stack test   # this should fail
$ cp examples/success-standard/src/FooBar.hs src/
$ stack test   # this should succeed
```

See https://github.com/commercialhaskell/stack-templates#introduction
for an introduction to the use of Stack .hsfiles templates.

Disclaimer: In Stack templates, `{{name}}` is what we call `<SLUG>` in
the v3 documentation, and the extra parameter `{{camel-name}}` is what
we call `<NAME>` in the v3 documentation. This is because `{{name}}` is
fixed.