[Elixir] Concept exercise: conditionals (#1129)
This introduces the concept exercise of conditions, similar to C#'s enums concept exercise, but pared down a little to limit to one conditional form and the use of atoms.  I think this would be an interesting starting point to use for an _ongoing narrative_ to build on.

And by working through this, I think it elicited a need for a `booleans` exercise to arise before this one.

New files introduced:
```text
.
├── .docs
│   ├── after.md
│   ├── hints.md
│   ├── instructions.md
│   └── introduction.md
├── .formatter.exs
├── .gitignore
├── lib
│   └── log_level.ex
├── .meta
│   ├── config.json
│   ├── design.md
│   └── example.ex
├── mix.exs
├── README.md
└── test
    ├── log_level_test.exs
    └── test_helper.exs
```

Also has some administrative updates for config, concepts, etc.