[JavaScript] Improve Concept Exercise: strings (#1353)
* js: change the story of the string concept exercise

The previous one need to implement a reversing task but this exercise is prior to array. So, we'll introduce String#trim instead.

* js: change instructions and hints

I changed instructions.md and hints.md according to the story's change.

* js: change stub and test files

I changed string.js, strings.spec.js and example.js according to the story's change.

* js: add contriutors to config.json

I added myself as a contributor of the string concept exercise.