Improvement/Changes to after document
Format changed so differs from introduction.

Took a risk and added documentation types for @doc and @moduledoc. I am
aware of the mention of @doc in the strings exercise, but I think the
mention of it here without specific reference to its underlying heredoc
string type could be useful and not interfere with the progression of
concepts.