[Docs] remove TODOs from Rust reference
Any work should be itemized as GitHub issue so we can delegate
appropriately and have an idea of what is left. Given our need to
complete more concept exercises I have removed this list. The general
tasks can be included as part of our contribution guide or itemized as
other issues.
Closes #1023.