[Rust] prototype vec concept exercise
I am changing my approach to creating concept exercises. In the past I
wrote the story and filled out the documentation to suit. That incurred
burdensome costs if the narrative didn't fit the Rust idioms in the
example.rs.
So I'll try creating the example code first for us to discuss. Based on
that progress I can fill out the exercism metadata.
(#1108)